[DF-2720] Removed inheritance from BaseObject at Component
Former-commit-id: 808f3a9bbaf0c8bcb70329cd9b21ca666a9136ca

+++ Tools/ResourceEditor/Classes/Qt/TextureBrowser/TextureCache.cpp
@@ -83,7 +83,7 @@ void TextureCache::setOriginal(const DAVA::TextureDescriptor *descriptor, DAVA::
 			while(iter.hasNext())
 			{
 				iter.next();
-				if(iter.value().weight < weightToRemove)
+				if(iter.value().weight < (size_t)weightToRemove)
 				{
 					cacheOriginal.remove(iter.key());
 				}
@@ -114,7 +114,7 @@ void TextureCache::setConverted(const DAVA::TextureDescriptor *descriptor, DAVA:
 			while(iter.hasNext())
 			{
 				iter.next();
-				if(iter.value().weight < weightToRemove)
+				if(iter.value().weight < (size_t)weightToRemove)
 				{
 					cacheConverted[gpu].remove(iter.key());
 				}
